test(negative): decouple stress operations from physical node operations
