+++ e2e/keywords/node.resource
@@ -66,15 +66,3 @@ Restart cluster
     FOR    ${statefulset}    IN    @{statefulset_list}
         wait_for_workload_pod_stable   ${statefulset}
     END
-
-Stress the CPU of all ${role} nodes
-    stress_node_cpu_by_role    ${role}
-
-Stress the CPU of all volume nodes
-    stress_node_cpu_by_volumes    ${volume_list}
-
-Stress the memory of all ${role} nodes
-    stress_node_memory_by_role    ${role}
-
-Stress the memory of all volume nodes
-    stress_node_memory_by_volumes    ${volume_list}
