+++ e2e/libs/keywords/node_keywords.py
@@ -1,10 +1,7 @@
 from robot.libraries.BuiltIn import BuiltIn
 
 from node import Node
-from node import Stress
 from node.utility import get_node_by_index
-from node.utility import list_node_names_by_role
-from node.utility import list_node_names_by_volumes
 
 from utility.utility import wait_for_all_instance_manager_running
 
@@ -13,7 +10,6 @@ class node_keywords:
 
     def __init__(self):
         self.node = Node()
-        self.stress = Stress()
 
     def reboot_volume_node(self, volume_name):
         volume_keywords = BuiltIn().get_library_instance('volume_keywords')
@@ -40,18 +36,3 @@ class node_keywords:
 
     def wait_for_all_instance_manager_running(self):
         wait_for_all_instance_manager_running()
-
-    def cleanup_stress_helper(self):
-        self.stress.cleanup()
-
-    def stress_node_cpu_by_role(self, role):
-        self.stress.cpu(list_node_names_by_role(role))
-
-    def stress_node_cpu_by_volumes(self, volume_names):
-        self.stress.cpu(list_node_names_by_volumes(volume_names))
-
-    def stress_node_memory_by_role(self, role):
-        self.stress.memory(list_node_names_by_role(role))
-
-    def stress_node_memory_by_volumes(self, volume_names):
-        self.stress.memory(list_node_names_by_volumes(volume_names))
